@@ -1,11 +1,8 @@
 package com.evil.helper.recycler.adapter;
 
-import android.content.Context;
-import android.support.annotation.LayoutRes;
+import androidx.annotation.LayoutRes;
 import android.view.View;
 import android.view.ViewGroup;
-
-import com.evil.helper.recycler.inface.OnAdapterItemClickListener;
 
 import java.util.List;
 
