@@ -1,8 +1,8 @@
 package com.evil.helper.recycler.holder;
 
 import android.content.Context;
-import android.support.annotation.IdRes;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.IdRes;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 /**
